Added ability for Dungeons to spawn Items
Added ability for Dungeons to hook into the developer's existing Items system by passing around Keys. The developer fills out the FDungeonItems DataTable with a unique key from which they can spawn an item from their system, and a name value that the Dungeon system will use. They choose the location and rotation of that item within a Room BP, then add the DungeonItem's name to the Room record. They also subclass the DungeonItemsHandler class and fill out the functions to cause any item passed to the HandleItem() function to properly enter their item system. It is assumed that the developer's item system will handle spawning, loading and unloading the items as well as saving and updating the item as appropriate, like any other items in their system.
As the PaperDungeon is generated it will decide whether that item should be spawned based on the SpawnLikelihood, and then call the DungeonItemsHandler->HandleItem() and the developer's code will take over.
Example items, Item DataTables, and BPs are provided.