Added testing class & unit testing
Split out the TestBuildDungeon() function into its own class, which now exercises the various rotations and functions in the DungeonBuilder.
It provides a quick visual check for developers to use to make sure rotation and offset are looking correct, and also provides unit tests for the rotation and offset functions.
Added ability to turn off Item generation.